destination-snowflake: rename InternalStagingSqlOp into StagingClient
